Java objects should own C++ State
Summary: When passing StateWrapper objects across the JNI, we were not ensuring that the Java objects would own the C++ state. This was initially done because I assumed that in Java, State would either be used immediately or discarded, so this wouldn't be unsafe. As it turns out, it makes sense in some cases to store the StateWrapper in Java and use it later, potentially even in other threads, so we need to make sure we maintain ownership of the C++ object from the Java object.

Reviewed By: shergin

Differential Revision: D15206194

fbshipit-source-id: a437d921ba00b194cf08bad80666bd99baf11d52