Consolidate issue templates (#24765)
Summary:
Eliminate the regression template, as it is hardly used and it is already served by the bug report template. Triagers can add the Regression label selectively, or we can do so through the bot.

Eliminate the discussion template, as it is not meant to be used and we can point people in the right direction using the existing questions template.

Rename the document bug template, and point people to the website repo.

Update the bug report template, and be less prescriptive about how an issue is reported. As long as the issue is well-described and the steps to repro can be reasonably followed by someone who is investigating the bug, we'll be happy.

[Internal] [Changed] - Consolidated issue templates.
Pull Request resolved: https://github.com/facebook/react-native/pull/24765

Differential Revision: D15276866

Pulled By: cpojer

fbshipit-source-id: d99a174982d2454d8cad3a195dfd627a07438611