Bring back JS bundle loading progress bar on Android
Summary: The progress bar on Android was disabled in D5329011 because of T19653381, but was never enabled again. I spent some time trying to reproduce the issue of the bundle being stuck while loading, but didn't succeed. So let's enable the progress bar and monitor whether people would start seeing this bug again.

Reviewed By: cpojer

Differential Revision: D17148134

fbshipit-source-id: 5130b809460bc743d26a6e88961f81061089fe1d